Initial Commit
The files for my bootcamp data analysis work. This commit features 2016 EU Referendum for the UK and 2011 UK Census data. My analysis mainly focuses on the 2016 EU Referendum in the UK dataset. Some analysis was done in Excel (2016 EU Referendum in the UK excel file) but more in depth in SQL (2016 Referendum Data and Census). I then also looked at the 2011 Census Data provided to us for this project and used it as a guideline to gather more insight into how the young people 20 - 35 (Millennials) voted in the UK in different areas.

As the dataset for the UK Census was in 2011, I did adjust the CSV file to account for the 5 years difference in the age groups. But this dataset is still only seen as a guide for the electorate population. It is not fully accurate to apply to the electorate population in 2016 as it won't take into consideration births, deaths and movement in population across different areas as well as missing some population statistics for a few areas featured in the Referendum.